Keys can be deleted
Many smallish improvements, the most noticable of which is the ability to delete keys through the key management page now (rather than throwing 500 errors like we used to). Cleaned up the LDAP connection settings a little bit more. Retested the LDAP functionality and confirmed it is still working.Settings page is still completely useless (partially due to the fact that ldap auth settings are set in the settings file, which is intended to be immutable- not 100% sure how to proceed with implementing this page yet).